// src/app/api/models/cronograma-request.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CronogramaRequest {
  activo?: boolean;
  anioLectivo?: number;
  fechaInicio?: string;
  interesMoraCuota?: number;
  montoCuota?: number;
  nivelEducativo?: 'PRIMARIA' | 'SECUNDARIA' | 'INICIAL';
  numCuotas?: number;
}