// src/app/api/models/cronograma-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface CronogramaResponse {
  activo?: boolean;
  anioLectivo?: number;
  fechaVencimiento?: string;
  id?: number;
  interesMoraCuota?: number;
  montoCuota?: number;
  nivelEducativo?: 'PRIMARIA' | 'SECUNDARIA' | 'INICIAL';
  numCuotas?: number;
}